tons of fixes

+++ src/WallpaperEngine/Data/Model/Object.h
@@ -527,7 +527,7 @@ struct ParticleData {
     std::string animationMode;
     float sequenceMultiplier;
     uint32_t maxCount;
-    uint32_t startTime;
+    float startTime;
     uint32_t flags;
 
     ModelUniquePtr material;
